+++ requestium/__init__.py
@@ -2,4 +2,4 @@ from selenium.common import exceptions  # noqa: F401
 from selenium.webdriver.common.keys import Keys  # noqa: F401
 from selenium.webdriver.support.ui import Select  # noqa: F401
 
-from .requestium import Session  # noqa: F401
+from .requestium_session import Session  # noqa: F401
